E5T4/features/frontendChanges/Stash#1
Changed the button list in this order:
Random, play, submit, previous, pause, next.
Added: when you click on the play button green color appears similarly with the Pause button the red color appears
Added: on the main page print all the clicks the program makes in the game board(matrix) to solve the game.